requirements.txt
Django==5.2.4
djangorestframework
psycopg2-binary